requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
behave-django==1.4.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.0.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-environ==0.11.2
django-extensions==3.2.3
django-modeltranslation==0.19.10
django-timezone-field==7.0
django-uuslug==2.0.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.4.1
h11==0.14.0
idna==3.10
kombu==5.4.2
outcome==1.3.0.post0
parse==1.20.2
parse_type==0.6.3
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
redis==5.1.1
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
text-unidecode==1.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
